Update mk-game-state so setting scene-id is a spec
Previously mk-game-state required a default scene-id as a required
argument which makes it less useful if you want to use the same function
for updating the game state multiple times for example to add new
enemies. The whole game state construction can now by represented by a
collection of specs.